+++ src/engine/link/useAbsolutePath.ts
@@ -1,9 +1,9 @@
-import React from 'react'
-import type { RouteKey, FileRoute, RouteStore } from '../route'
+import { createContext, useContext } from 'react'
+import type { FileRoute, RouteStore } from '../route'
 
-export const GlobalRouteContext = React.createContext<RouteStore>({})
+export const GlobalRouteContext = createContext<RouteStore>({})
 export const useAbsolutePath = ({ key }: Pick<FileRoute, 'key'>) => {
-  const globalRouteStore = React.useContext(GlobalRouteContext)
+  const globalRouteStore = useContext(GlobalRouteContext)
 
   return [...globalRouteStore[key]]!
 }
