--- src/pages/effects/index.ts
@@ -1,1 +1,0 @@
-export { EffectPage } from './instancePage'
