--- src/pages/moves/index.ts
@@ -1,2 +1,0 @@
-export { MovePage } from './instancePage'
-export { MovesIndexPage } from './indexPage'
