--- src/pages/spec/index.ts
@@ -1,1 +1,0 @@
-export { SpecPage } from './indexPage'
