+++ src/pages/spec/index.tsx
@@ -1,0 +1,7 @@
+import { createRoute } from '@src/engine/route/dirRoute'
+
+import { SpecPage } from './indexPage'
+
+export const specsRoute = createRoute({
+  index: <SpecPage />,
+})
