+++ src/pages/spec/indexPage.tsx
@@ -1,6 +1,3 @@
-import { createRoute } from '@src/engine/route/dirRoute'
-import React from 'react'
-
 export const SpecPage: React.FC = () => (
   <html>
     <head>
@@ -14,7 +11,3 @@ export const SpecPage: React.FC = () => (
     </body>
   </html>
 )
-
-export const specsRoute = createRoute({
-  index: <SpecPage />,
-})
